(Login) fix cookie session encoding bug
Turns out the cookie session management simply serializes the session
state by base64 encoding the data and sending that as the cookie.
When the encoded cookie size got too large the brower would simply
throw it away, thus destroying the session. This fix ensures that
at login we only store the userId in the session state, not the user
model as previously.